requirements.txt
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.3.0
spams==2.6.5.4
spams_bin==2.6.4
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
